Implement CanvasApplication class (#1051)
